+++ EchoMMO-Backend/src/main/java/com/echommo/dto/CreateListingRequest.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 
 @Data
 public class CreateListingRequest {
-    private Long userItemId; // [FIX] Đổi Integer -> Long để khớp với UserItem mới
+    private Integer userItemId; // [FIX] Dùng Integer
     private Integer quantity;
     private BigDecimal price;
 }
