+++ EchoMMO-Backend/src/main/java/com/echommo/service/InventoryService.java
@@ -1,3 +1,5 @@
+// File: EchoMMO-Backend/src/main/java/com/echommo/service/InventoryService.java
+
 package com.echommo.service;
 
 import com.echommo.entity.User;
@@ -6,12 +8,15 @@ import java.util.List;
 
 public interface InventoryService {
     List<UserItem> getInventory(Integer charId);
-    void equipItem(Integer charId, Long userItemId);
-    void unequipItem(Integer charId, Long userItemId);
-    UserItem enhanceItem(Integer charId, Long userItemId);
+
+    // [FIX] Đổi Long -> Integer để chiều lòng client
+    void equipItem(Integer charId, Integer userItemId);
+    void unequipItem(Integer charId, Integer userItemId);
+    UserItem enhanceItem(Integer charId, Integer userItemId);
+
     User expandInventory(User user);
     void addItemToInventory(User user, Integer itemId, int quantity);
 
-    // [NEW] API SỬA ĐỒ
-    UserItem repairItem(User user, Long userItemId);
+    // [FIX] Đổi Long -> Integer
+    UserItem repairItem(User user, Integer userItemId);
 }
